refactor: lotto 패키지내 number, rank 패키지 분리
- lotto 패키지 내 Number 관련 클래스는 lotto.number 패키지로 이동
- Rank 클래스는 Rank 패키지로 분리
