+++ README.md
@@ -25,13 +25,13 @@
 - [x] 구매한 로또 번호를 오름차순으로 정렬한다.
 - [ ] 입력받은 보너스 번호를 통해 당첨 보너스 번호 생성한다.
 - [ ] 주어진 수만큼 로또를 생성을 반복한다.
-- [ ] 구매한 로또 번호와 당첨 번호를 비교한다.
-- [ ] 보너스 번호가 포함되어 있는지 확인한다.
-- [ ] 보너스 번호가 포함되어 있지 않고, 6개의 숫자가 일치하면 1등이다.
-- [ ] 보너스 번호가 포함되어 있고 5개의 숫자가 일치하면 2등이다.
-- [ ] 보너스 번호가 포함되어 있지 않고, 5개 숫자가 일치하면 3등이다.
-- [ ] 보너스 번호가 포함되어 있지 않고, 4개 숫자가 일치하면 4등이다.
-- [ ] 보너스 번호가 포함되어 있지 않고, 3개 숫자가 일치하면 5등이다.
-- [ ] 등수에 따라 당첨금을 계산한다.
+- [x] 구매한 로또 번호와 당첨 번호를 비교한다.
+- [x] 보너스 번호가 포함되어 있는지 확인한다.
+- [x] 보너스 번호가 포함되어 있지 않고, 6개의 숫자가 일치하면 1등이다.
+- [x] 보너스 번호가 포함되어 있고 5개의 숫자가 일치하면 2등이다.
+- [x] 보너스 번호가 포함되어 있지 않고, 5개 숫자가 일치하면 3등이다.
+- [x] 보너스 번호가 포함되어 있지 않고, 4개 숫자가 일치하면 4등이다.
+- [x] 보너스 번호가 포함되어 있지 않고, 3개 숫자가 일치하면 5등이다.
+- [x] 등수에 따라 당첨금을 계산한다.
 - [ ] 당첨된 로또의 총 수입금을 계산한다.
 - [ ] 수익률을 계산한다.
