+++ src/main/kotlin/domain/lotto/PurchasedLotto.kt
@@ -1,5 +1,9 @@
 package domain.lotto
 
+import domain.lotto.number.BonusNumber
+import domain.lotto.number.LottoNumber
+import domain.rank.Rank
+
 class PurchasedLotto(lottoNumbers: List<LottoNumber>) : Lotto(lottoNumbers) {
     fun getSortedLotto(): Lotto = Lotto(this.sortedBy { it.value })
