+++ src/main/kotlin/domain/lotto/number/BonusNumber.kt
@@ -1,3 +1,3 @@
-package domain.lotto
+package domain.lotto.number
 
 class BonusNumber(value: Int) : LottoNumber(value)
