+++ src/main/kotlin/domain/rank/Rank.kt
@@ -1,4 +1,4 @@
-package domain.lotto
+package domain.rank
 
 enum class Rank(val countOfMatch: Int, val winningMoney: Int) {
     FIRST(6, 2_000_000_000),
