+++ src/test/kotlin/domain/lotto/LottoTest.kt
@@ -1,5 +1,6 @@
 package domain.lotto
 
+import domain.lotto.number.LottoNumber
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
