+++ src/test/kotlin/domain/lotto/WinningLottoTest.kt
@@ -1,5 +1,7 @@
 package domain.lotto
 
+import domain.lotto.number.BonusNumber
+import domain.lotto.number.LottoNumber
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
