+++ src/test/kotlin/domain/lotto/number/LottoNumberTest.kt
@@ -1,5 +1,6 @@
-package domain.lotto
+package domain.lotto.number
 
+import domain.lotto.number.LottoNumber
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
