+++ src/test/kotlin/domain/rank/RankTest.kt
@@ -1,4 +1,4 @@
-package domain.lotto
+package domain.rank
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
